requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Telegram Bot
python-telegram-bot==20.7

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Settings & Environment
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Gemini AI
google-generativeai==0.8.0

# Image Processing
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# HTTP Client (httpx versiyasi python-telegram-bot bilan mos bo'lishi kerak)
httpx~=0.25.2
aiofiles==23.2.1

# Utilities
python-dateutil>=2.9.0
